backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
pydantic==2.5.3
python-multipart==0.0.6
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
regex==2023.12.25
google-generativeai==0.3.1
authlib==1.2.1
python-jose[cryptography]==3.3.0
httpx==0.25.0
pymongo==4.6.1
motor==3.3.2
